Update universite-du-quebec-a-montreal.csl
I made changes to answer as much as possible to your comments. I tried to remove as much duplications as possible. UQAM style is quite complex and manage a large number of different documents. There are a lot of differences in punctuation that can explain why it is often hard to merge several types of documents under the same rules. It is also written to work with both Zotero and Mendeley. This is why sometimes it could look weird and is longer because Mendeley has less document types and less fields.
Your comments are always welcome.