Update civitas-journal-of-social-sciences-abnt.csl
The editors are planning to change family-names from "uppercase" to "capitalize-first". I have include the new format in the first version of the style. Now the option is to maintain "uppercase" while format change is not implemented. Also 'delimiter-precedes-et-al="never"' was included here. In Social Sciences this is not necessary for the other "contributors".
I simplified the cases for "author" (no more "if-cases").
"Entry-encyclopedia" became part of "locators".
In bibliography I have included "speech", the variable "genre" for "manuscripts", and I replaced "entry-encyclopedia" (putting together with "chapter" etc.).