Create gb-t-7714-2015-numeric-zh-cn.csl
"GB/T 7714-2015" is a national standard/guidelines of the People's Republic of China for bibliographic references and citations to information resources. It is widely used in academic journals and papers in mainland China. It is drafted with reference to the international standard ISO 690:2010(E) and its previous version is "GB/T 7714-2005".