Update university-of-bradford-harvard.csl (#4127)
Changed the following after further revision and discussion with a teacher to fix the below as per: https://www.bradford.ac.uk/library/find-out-about/referencing/Guide-to-referencing-using-the-Harvard-System--November-2017.pdf

Inline citation:

removed page number(s)
Bibliography:

The ", " delimiter after each group was not correct. Replaced with " " and added ". " after titles and ", " after volume/issue of journals.
Removed the "p./pp." label before page number(s).
Changed the format of place and publisher as "Place: Publisher".
Set the space between bibliography lines to "0" as per the EndNote style provided by the university.
Set italic for the title if publication is post or post-weblog.
Fixed the separator between publisher and pages for chapters.
Fixed date and url order for post post-weblog, and websites.
Reverted back to "Accessed" for access date on post post-weblog, and websites. (There are contrasting directions between the brief citation guide and the full reference. The guidelines from the full reference was used here as it's the same used on EndNote's style from the university)
Added page locator (if present) to inline citations.