Update zilsel.csl
Thank you,
I confirm this is a “note” type citation style, my mistake. This style is intended both for French and English citation (considering that the papers published in this journal are all in French language so far).  I add a “template” reference to the style I used. Unfortunately the journal doesn't have yet an online description of the style.