Update harvard-university-for-the-creative-arts.csl
Sorry to open a new pull request but when this was closed the other day when an edited book (not chapter) was referenced in a bibliography the ed. or eds. after the name was no longer within brackets.
It is producing this: 
Hancké, B. et al. eds. (2007) Beyond varieties of capitalism : Conflict, contradiction, and complementarities in the European economy. Oxford and New York: Oxford University Press.
When it should be this:
Hancké, B. et al. (eds.) (2007) Beyond varieties of capitalism : Conflict, contradiction, and complementarities in the European economy. Oxford and New York: Oxford University Press.
I use the visual editor and when I amended the names variable to the macro editor it fixed this issue.
Is this change ok, or is there a different way that it should be done?