Create apa-7th-edition-with-comments.csl
This is the same apa.csl but with the comments option, as is illustrated in the book of the APA Style 7th edition. This contribution can be helpful for the use of the APA Style indicating the original source of translation of books, articles, and so on, at the and of the reference entry.