Create the-university-of-tokyo-law-review.csl
Hi, dear Committers.
I'm a postgraduate student in Japan, majoring in law.
I'm writing master's thesis by using software tool, but unfortunately I can't find the CSL I want from the Internet.
So I wrote the CSL. I believe this CSL is helpful for students and researchers in law field in Japan.
This CSL keeps The The University of Tokyo Law Review Rules. This is the most authoritative university law review in Japan.
I hope you adopt this, and reach to another law writers in Japanese.

Regards,
ysmatsud a.k.a. MATSUDA Yoshiro