Create china-national-standard-gb-t-7714-2015-author-date.csl
There is already an old author-date version for China National Standard GB/T 7714 in the repository, named "chinese-gb7714-2005-author-date.csl". 
However，The latest version of the China standard “GB/T 7714” is “GB/T 7714-2015”. This version, "china-national-standard-gb-t-7714-2015-author-date.csl", is keeping the CSL's support to “GB/T 7714” up-to-date with the latest version.