Create the-open-university-harvard-s390-modules.csl
This style is based on "The Open University - Harvard" (the-open-university-harvard.csl) style and has been adjusted to meet the specific requirements of the S390 project modules. Specifically, the in-text citations are limited to 3 authors max, with 'et-al' in italic font. In the bibliography, the authors are limited to 10 authors and 'et-al' in italic font.  Credit is given to original authors through the "template" reference as per the style requirements here: https://github.com/citation-style-language/styles/wiki/Style-Requirements.