Create turabian-9-author-date
As far as I know, the Turabian 9th edition is very similar to Chicago 17th. The only difference I am aware of for the author-date variant is using "no. x" for journal issue numbers.
So instead of: "Random Journal 5 (1): 123-321."
You will see: "Random Journal 5, no. 1: 123-321."

My PhD program requires Turabian 9th edition, author-date. So here is the new file, which is based off Chicago 17th. If there are other differences, I am not aware of it. If others can modify this file, great! If not, this addition is still better than nothing. The last Turabian author-date variant we have in the Zotero repository is in 2013.