research-institute-for-nature-...: drop 'in' if no editor *
This allows for a bit better compatibility with the
distinction between InBook and InCollection after importing
a bib file into Zotero.
Previously, a InBook with a series caused
'In: <Series>, <Publisher>, p. x-y' while the series is not
the container (a container-title, i.e. booktitle in bib, is
missing in case of InBook). Dropping 'In:' in that case makes
this a bit nicer.