research-institute-for-nature-...: cases with container-title *
These cases are now better handled.
E.g. this makes it better appropriate to distinguish between
cases of type 'chapter' with (InCollection) and without (InBook)
container-title.
The presence of a container-title, except for articles, now
invokes the use of "In:" regardless of the presence/absence of
editor/translator (hence relaxing e7cc2c2).
Also, it invokes putting the edition after the container-title.