Update apa-no.csl
I have changed the title, short title, link to self and template. Instead of setting the documentation link to APA I have changed it to the home page for the Norwegian APA manual. Please advise if more changes are needed or if some should be retracted. BTW, the template link in APA 7 is to this address:  <link href="http://www.zotero.org/styles/apa-6th-edition" rel="template"/>