Create ypfs-kpd.csl
First iteration of Yale Program on Financial Stability's Key Program Document style. Validated by CSL Validator. This style allows contributors to the [Journal of Financial Crises](https://elischolar.library.yale.edu/journal-of-financial-crises/) quickly format resources for the Key Program Documents section.
It uses stock in-text citations from Chicago 17 (author-date) but creates a new bibliographic entry: emphasizing the title of the resources; linking the in-text citation with the reference; providing the month and day when available (much of the Journal's subjects operate on timescales of days, weeks, or months), a short description of the resource, and a link to the resource in YPFS' [Resource Library](ypfs.som.yale.edu). This style leverages two never-used fields in Chicago 17 (title-short and call-number), allowing an author to use both in the same document.