Update copernicus-publications.csl
I tried to get some answers related to this style in #5224, but no one responded. 
I deleted type="software" which I had originally included, to be compliant with 1.0.1 (the validator did not check out otherwise).
Also, for type="dataset" is chose to refer to the repository in the field "archive"
There has been a discussion about this here https://github.com/citation-style-language/styles/issues/4367 
but to no end. 
Otherwise, this now reproduces the style for all journals by Copernicus Publications.