Update society-of-biblical-literature-fullnote-bibliography.csl
When a custom citation is triggered by the annote variable being used, this change inserts a comma and space (rather than space only) as a delimiter before the locator. This helps prevent the need to store the comma in annote and improves interoperability with similar styles (e.g., Catholic Biblical Association) that require an identical abbreviated reference but a different delimiter (space only).