Update university-of-roehampton-harvard.csl
Apologies for the double-posting; I couldn't work out how to modify my existing change proposal!

When generating citations, it was omitting the 'et al' on resources with 4+ authors. I've added it to line 449 this time (which I hope is the correct one), but can't be sure as I was ultimately using the Visual CSL editor to put it together in the first instance.