Update chicago-author-date.csl
Some changes proposal that will improve the Chicago author-date style in portuguese language. (both portuguese-portugal and portuguese-brasil)

It would be great to translate the following:
In line 628: <text value="filed"/>
and line 638  <text value="issued"/>
It would be better to have instead of filed - submetida em
and instead of issued - publicada em
But I guess it's not possible to localize text values, just terms.


Also I would like to propose a change for the all the languages. This changing is related how the volume in journal articles is shown. 
In this style code the volume number is shown as:   40 (3)
And the Chicago Manual of Style states that has to be 40, no.3

As stated in: https://www.chicagomanualofstyle.org/book/ed17/part2/ch09/psec027.html

And so, I would suggest the following alteration:
   
<if type="article-journal">
        <choose>
          <if variable="volume">
            <text variable="volume" prefix=" "/>
            <group prefix=", ">
              <choose>
                <if variable="issue">
                  <text variable="issue" prefix="no."/>