Update deutsche-gesellschaft-fur-psychologie.csl
I needed this citation style to work with translated works, which was not implemented in the current version. I added this functionality to the style for Deutsche Gesellschaft für Psychologie.

I also took the liberty to update some references, since the link to the 4th edition was not working. I cannot tell if the citation rules for translated books or classic works are the same in the 4th edition. The rest of the citation style is compatible and compliant with the 5th edition, so I assume this was safe to update.

I have impelemented two cases:
1) classic works
2) translated works

They are being cited differently and the bibliography entry looks slightly different for these two cases. I decided to use the "translator" variable to distignuisch the two cases, I assumed this would be safe.