Fix the order for conference papers in APA styles
The current code in the macro date-sort-date is outputting in all
cases the same
   <date variable="issued" form="numeric"/>
except for conference papers with editors etc., where not output
at all is given. Consequently these conference papers will be moved
to the end of a block or the bibliography at all (apa-cv showed this
error instructively. This fixes the macro as discussed in
https://forums.zotero.org/discussion/94197/ .

It updates all styles which where be found with
> grep 'macro name="date-sort-date"' *.csl