Update pontificia-universita-gregoriana-teologia.csl
That's good! I tried to write a code that converts the roman numbers in arabical numbers, but I see that is not working. Now, the volume macro works fine for the multiple-volume books. You know if there are some method to convert the roman numbers in arabical numbers; because the gregorian stile only accept the arabical numbers. Thanks!

Co-authored-by: anidal <avidalgomez@gmail.com>