Update pontificia-universita-gregoriana-teologia.csl
Lines 303 and 322: for the reviewed articles, the "editor" is changed with the "reviewed-author". Norme Tipografiche p.35, lines 995-1010
Line 183: The volume variable admits numbers in roman. That is pendent of revision.
Line 475: The number of authors on the bibliografy for each citation is limited to 4, and for the books with more than 4 authors, it is limited to 2. Norme tipografiche p.34 lines 960-980