Update apa.csl
This is for localizing the “Original work published” text that appears at the end of a reference when the original date of a work is present. I created the term "original-work-published" within each locale section. Then, I used Google Translate to add a translation to each locale (but for Spanish, which is my first language). Next, in the macro “publication history", I replaced the “text value” with the correspondent “text term” I created. 
I also corrected some mistakes in Latin (la) and Lithuanian (lt) locales.