Fix malformed XML in raptor-journal.csl
According the XML specification, the `<?xml ?>` declaration must be at the beginning of an XML file and nothing should precede it.

I encountered this error when I tried to read the file using citeproc-java. It seems the Java XML parser does not like byte-order marks. I hope, the BOM was not necessary in this case but it seems the file is just normal UTF-8.