Create vancouver-brackets-doi-no-et-al.csl
This is a tweak of the vancouver-brackets-no-et-al style:

68ed5f7645a5238ee7e62e7540dcada514fd1c24

The tweak is removing the "available at" URL and "accessed-date" elements for records containing a DOI field. Instead, the DOI is displayed in the "doi: 0000000/000000000000" notation as per documentation:

https://www.nlm.nih.gov/bsd/uniform_requirements.html#electronic

This tweak is especially important for the health and medical field, and I am going to upload a dependent NLM style next. As the foreword to the last edition of the NLM Style Guide put it:

> For materials on the Internet the medium designator [Internet] is required following the title, and the URL is required. In 2015, almost everyone recognizes a string starting with http: or https: or doi: as a URL or address for something on the Internet. Do we still need to require Internet in brackets as a medium designator?

Source: https://www.ncbi.nlm.nih.gov/books/NBK310495/