Create university-of-bucharest-faculty-of-law
Good afternoon,

My name is Alexandru Panait and I would like to add this style to the Zotero Style Repository. This is the citation and bookmark style that University of Bucharest Faculty of Law requests to be followed in the writing of the Dissertation Work in all of its Masters Program in year 2023-2024.
Please note that I am a student that was asked by a teacher to create this code (unofficially, because our university does not care of its students and would never direct resources in order to create another code to make our lives easier by using Zotero. Also from experience I can tell you that they will tolerate the situation.) for the sole reason that I remember some html coding from highschool. The code is good (I hope), I followed the steps and I introduced the code in the Formatter engine and Validation engine.