Create university-of-bucharest-faculty-of-law.cls
This is the style of citation and bibliography requested to be followed by the University of Bucharest Faculty of Law in creation of Dissertation Work for all its Masters programs. It is the same since 2018.
I would like to contribute with this code in the enlargement of Zotero Style Repository.