Create national-library-of-medicine-grant-proposals-all-names-doi-pmid-pmcid
There are people who prefer to list all of DOI, PMID, and PMCID at the end of each reference entry. NIH grants also forbids URL in bibliography, and some papers would show up URLs as source.

This is based on national-library-of-medicine-grant-proposals. I'm also happy to have this merged to that one, but I would appreciate helps (this is my first time doing GitHub PRs and creating Zotero styles)