Create university-of-bucharest-faculty-of-law.csl (#6791)
* Create university-of-bucharest-faculty-of-law.csl

This is the requested style of citation and bibliography to be followed by students of University of Bucharest Faculty of Law in creating their Dissertation Work for all Masters programs (style used since 2018).
I would like to contribute to the enlargement of Zotero Style Repository with this html code.
Please note that I am a student that was asked by a teacher to create this html code for the sole reason that I remember some html coding from highschool.
I created this code unofficially, because our University does not care about the digitalization process and the creation of tools in order to help students. They will not direct resources in creating another code and probably will tolerate the situation. I will speak with the Vice Dean about the existence of this code that will be added to Zotero Style Repository.

* Fix basics

* fix self link

* Removed the contributor from `summary`

* Update university-of-bucharest-faculty-of-law.csl

---------

Co-authored-by: Sebastian Karcher <karcher@u.northwestern.edu>
Co-authored-by: Brenton M. Wiernik <bwiernik@users.noreply.github.com>