Allow listing of a source database
See ¶ 14.72. While `source` might be the obvious field to print for this (since Zotero imports the names of library catalogues here), this would result in the names of databases appearing after many entries for Zotero users where they are not desired, since it excludes URLs from the citation data by default with page numbers. The suggestion of @bwiernik to use `archive` follows the CSL styles for APA and MLA.