Updates to USGS citation style to conform to "USGS Suggestions to Authors 8th edition" (#7477)
* Further updates for conformity to USGS Suggestions to Authors 8th edition.

Reference styles are provided in "Suggestions to Authors of the Reports of
the U.S. Geological Survey—The Official Style Guide of the U.S. Geological
Survey, 8th Edition, Revised". As of March 2025, this publication is an
internal USGS document. I added format information for common publication
types in comment at top of file.

- Add dataset and software types.
- Add DOIs to journal articles and reports using URL with accessed date.
- Report: Add option to use genre.
- Webpage: Move institution to proper location.

* Remove obsolete hack using book with version instead of software.

* Fix CSL so layout is much more robust to missing fields.

* Re-indent CSL styles

---------

Co-authored-by: baagaard-usgs <22621443+baagaard-usgs@users.noreply.github.com>