Adding style for L'Homme *for English-language manuscripts*
This style is based on the style for L'Homme already present in the repo. The reason for producing an English-language version is that the journal accepts manuscripts written in English, and the reference stylesheet needed to be adapted accordingly (using "and" instead of "et", and so on).
I have checked it locally and it works (but I'm happy to chisel further if useful).