Remove `generic-base` category from derivatives
According to the CSL spec, `generic-base` is for 'used for generic styles like Harvard and APA', but numerous local derivatives for use with specific journals and publishers have retained it. This removes the category from such files, so that it is possible again to browse the actual generic styles on the Style Repository.