Create bluebook-law-review-with-abstract.csl
This is based on the latest version of bluebook-law-review but adds one line of code that will include the abstract field after each citation (<text variable="abstract" display="block"/>). This is based on what was done with apa-with-abstract. If there's any way to automatically or semiautomatically update this dependent style when the main style is updated or a different way this should be submitted to make that happen, let me know.