set to 3/1 et al.
While the guidelines would want 4/1 ("Use “et al.” for citations with four or more authors: Michler, Tjernström, Verkaart, and Mausch
(2019) should be cited as Michler et al. (2019)."), I don't see any paper that actually follows that and they just use the classic 3/1 rules. Given this is a Wiley journal which are typically bad with keeping their guidelines up to date, I switched this to 3/1.

Examples: 
https://onlinelibrary.wiley.com/doi/epdf/10.1111/ajae.12534
https://onlinelibrary.wiley.com/doi/epdf/10.1111/ajae.12472