Update harvard-university-for-the-creative-arts.csl (#7346)
* Update harvard-university-for-the-creative-arts.csl

This change is to do with Directors in the CSL. Zotero is using CSL 1.0.2 and uses the Director variable and tag, this doesn't display at the moment in our current CSL when referencing with Zotero. I made a commit in the summer but withdrew it as the changes suggested by the reviewer had a knock on impact, as I've since discovered that other software that students use at UCA (Paperpile, BibGuru, myBib) don't currently use CSL 1.0.2. and won't display the director or tag if I update the code as was suggested.

Would it be possible to approve this commit, where the code will display either director or author and use a prefix instead of a label?

"The "Director" name variable was introduced last year with CSL 1.0.2. Therefore, a lot of citation styles don't include it yet."

From <https://forums.zotero.org/discussion/comment/476797/#Comment_476797>

* Update harvard-university-for-the-creative-arts.csl

I received an error code before and was told that something was down, so just trying the same code again - hope that is ok