Create metropolia-university-of-applied-sciences-english.csl
This style is for the Metropolia University of Applied Sciences in Finland. It will be used by students in the EBA and IBL degree programs. I am the senior lecturer responsible for teaching this style guide to students when they write their theses, but the students will use this style guide in most of their courses.