Update japan-journal-of-stuttering-and-other-fluency-disorders.csl
"author-yomi", which cannot be validated, is replaced by original-author. This may cause a conflict with the native ussage of the variable original-author, but it is expected to be very infrequent. However, the user should be warned about it. Hopefully this change would result in successful validation. There are other small changes like updated summary update that now contain both Japanese and English summaries in one place.