Chicago: Cite magazines by date only
Cite magazines by their full date only, using the same formatting as newspapers (but with the option of page ranges). See _CMOS_ 14.87: 'Weekly or monthly (or bimonthly) magazines, even if numbered by volume and issue, are usually cited by date only.' Previously, the styles interpreted 'usually' as meaning that volume or issue numbers could be printed if available. _CMOS_ staff have clarified that the example of a magazine cited with a volume number in 14.100 is a holdover from the 15th edition that has been overlooked in subsequent revisions. In 14.65, there is the option of applying 'journal form' to magazines, but _CMOS_ staff recommend that this should only be applied to items that the author feels is best presented as part of a journal, such as a periodical with articles that give formal citations; an item treated as a magazine should always be cited by full date only.