Update council-of-science-editors-author-date.csl (sort alphabetically)
via https://forums.zotero.org/discussion/126699/style-error-cse-name-year-9th-edition/p1

[Guidelines](https://www.csemanual.org/Tools/CSE-Citation-Quick-Guide.html): "Name–Year

The following examples illustrate the name–year system. In this system, in-text references consist of the surname of the author or authors and the year of publication of the document. End references are unnumbered and appear in alphabetical order by author and year of publication, with multiple works by the same author listed in chronological order."