Chicago: Move locator with archival citations
When providing a physical locator (`column`, `folio`, `page`) with an item that has a full archival reference (at least `archive` and `archive_location`), assume that the locator is part of the archival item (as shown in CMOS18 14.127). The locator must always be labelled in this case.

Fix a regression introduced with the styles for the 18th edition wherein users could receive an archival reference formatted as a database reference (as in CMOS18 14.113). The styles now revert to following the CSL specification in always handling `archive_location` as a 'storage location', rather than attempting to use this as a database reference in some situations, which is now assumed to be stored in `number`. This corrects the display of items where the archive place is part of the archive name, as in 'Lambeth Palace Library, MS 51'.

In shortened notes, when citing multiple items with a `genre` only (such as untitled introductions by the same author), disambiguate using a `container-title` if necessary.