MLA: Improve container and event handling
Follow the MLA Handbook more closely in using the prescribed positions for events and archival references in Container 1.

Do not substitute a `publisher` for an author name with software: the MLA Handbook provides an example of this, but this seems to have been an error, as it has been removed in the online version.